DjangoService/requirements.txt
Django==5.2.1
djangorestframework==3.15.0
django-cors-headers==4.3.1
django-filter==24.1
mysqlclient==2.2.4
Markdown==3.6
coreapi==2.3.3
python-dotenv==1.0.1 